requirements.txt
opencv-python==4.9.0.80
python-dotenv==1.0.1
python-ffmpeg==2.0.12
python-socketio==5.11.4
torch==2.2.1
torchvision==0.17.1
ultralytics==8.2.95
websocket-client==1.8.0

numpy==1.24.3
mtcnn==0.1.0
keras-facenet==0.3.2
pickle-mixin==1.0.2
tensorflow==2.12.0
scipy==1.11.3
scikit-learn==1.3.0





# numpy
# mtcnn
# keras-facenet
# pickle-mixin
# tensorflow
# scipy
# scikit-learn